+++ src/one_layer_word_fashion_embedding.py
@@ -2,9 +2,9 @@ from tensorflow.keras.datasets import fashion_mnist
 from sklearn.model_selection import train_test_split
 from sklearn.metrics import classification_report
 from sklearn.metrics import confusion_matrix
+from scipy.spatial.distance import cosine
 
-from generator_models import TextEncoderGeneratorModelCreator
-from generator_models import TextDecoderGeneratorModelCreator
+from generator_models import GeneratorModelCreator
 from discriminator_models import DiscriminatorModelCreator
 from gan_models import EncoderGanModelCreator
 from gan_models import DecoderGanModelCreator
@@ -20,13 +20,14 @@ from numpy import ones
 from numpy import zeros
 import numpy as np
 import re
+import os
 import constants
 
 import sys
 
 """
 One layer encoding-decoding:
-Text -> (Encode) -> Fashion -> (Decode) -> Text
+Word -> (Encode) -> Fashion -> (Decode) -> Word
 """
 
 """
@@ -78,24 +79,50 @@ sorted_words_tuple = sorted_words_tuple[:100]
 sorted_words = [tuple[0] for tuple in sorted_words_tuple]
 
 words = [word for word in words if word in sorted_words]
-
-# Tokenize text data
-tokenizer = DefaultTokenizer()
-_, _, word_index, index_word, vocabulary_size = tokenizer.tokenize_corpus(
-    words)
-
-tokens = [word_index[word] for word in words]
-tokens = np.asarray(tokens)
-
-# Over sampling the minority data
 oversample = RandomOverSampler(sampling_strategy='minority')
-tokens = tokens.reshape(-1, 1)
-tokens, tokens = oversample.fit_resample(tokens, tokens)
+words = np.asarray(words)
+print('Words shape before over sampling: {}'.format(words.shape))
 
-tokens_train, tokens_test = train_test_split(tokens,
-                                             test_size=0.15)
-print('Tokens train shape: {}'.format(tokens_train.shape))
-print('Tokens test shape: {}'.format(tokens_test.shape))
+words = words.reshape(-1, 1)
+words, words = oversample.fit_resample(words, words)
+print('Words shape after over sampling: {}'.format(words.shape))
+
+# Convert to word embedding
+original_embeddings_index = {}
+with open('data/glove.6B.50d.txt') as f:
+    for line in f:
+        values = line.split()
+        word = values[0]
+        coefs = np.asarray(values[1:], dtype='float32')
+        original_embeddings_index[word] = coefs
+print('Found %s word vectors.' % len(original_embeddings_index))
+
+embeddings_index = {}
+word_embeddings = []
+for word in words:
+    embedding_vector = original_embeddings_index.get(word)
+    if embedding_vector is not None:
+        embeddings_index[word] = embedding_vector
+        word_embeddings.append(embedding_vector)
+    else:
+        print('Word not found: {}'.format(word))
+
+word_embeddings = np.asarray(word_embeddings)
+print(word_embeddings.shape)
+
+# Create train and test set
+words_train, words_test = train_test_split(words,
+                                           test_size=0.15,
+                                           random_state=1)
+word_embeddings_train, word_embeddings_test = train_test_split(word_embeddings,
+                                                               test_size=0.15,
+                                                               random_state=1)
+
+print('Train shape: {}'.format(words_train.shape))
+print('Test shape: {}'.format(words_test.shape))
+
+print('Embeddings train shape: {}'.format(word_embeddings_train.shape))
+print('Embeddings test shape: {}'.format(word_embeddings_test.shape))
 
 # Load fashion data
 (fashion_image_train, _), _ = fashion_mnist.load_data()
@@ -110,11 +137,12 @@ Create models
 
 """ Encoder """
 
-# text -> image
-text_encoder_creator = TextEncoderGeneratorModelCreator(constants.INPUT_SHAPE,
-                                                        constants.OUTPUT_SHAPE,
-                                                        vocabulary_size)
-text_encoder = text_encoder_creator.create_model()
+# word -> image
+word_encoder_creator = GeneratorModelCreator(input_shape=(50,),
+                                             output_shape=constants.OUTPUT_SHAPE,
+                                             to_image=True,
+                                             activation='tanh')
+word_encoder = word_encoder_creator.create_model()
 
 # Discriminator
 encoder_discriminator_creator = DiscriminatorModelCreator(
@@ -122,21 +150,21 @@ encoder_discriminator_creator = DiscriminatorModelCreator(
 encoder_discriminator = encoder_discriminator_creator.create_model()
 
 # GAN (Combine encoder generator and discriminator)
-encoder_gan_creator = EncoderGanModelCreator(text_encoder,
+encoder_gan_creator = EncoderGanModelCreator(word_encoder,
                                              encoder_discriminator)
 encoder_gan = encoder_gan_creator.create_model()
 
 """ Decoder """
 
-# image -> text
-text_decoder_creator = TextDecoderGeneratorModelCreator(constants.INPUT_SHAPE,
-                                                        vocabulary_size)
-text_decoder = text_decoder_creator.create_model()
+# image -> word
+word_decoder_creator = GeneratorModelCreator(input_shape=constants.INPUT_SHAPE,
+                                             output_shape=50,
+                                             from_image=True)
+word_decoder = word_decoder_creator.create_model()
 
 # GAN (Combine state2image generator and image2state generator)
-decoder_gan_creator = DecoderGanModelCreator(text_encoder,
-                                             text_decoder,
-                                             loss='sparse_categorical_crossentropy')
+decoder_gan_creator = DecoderGanModelCreator(word_encoder,
+                                             word_decoder)
 decoder_gan = decoder_gan_creator.create_model()
 
 """
@@ -145,8 +173,8 @@ Create trainers
 
 """ Encoder """
 
-# text -> image
-encoder_trainer = EncoderTrainer(text_encoder,
+# word -> image
+encoder_trainer = EncoderTrainer(word_encoder,
                                  encoder_discriminator,
                                  encoder_gan,
                                  training_epochs=constants.TRAINING_EPOCHS,
@@ -154,9 +182,9 @@ encoder_trainer = EncoderTrainer(text_encoder,
 
 """ Decoder """
 
-# image -> text
-decoder_trainer = DecoderTrainer(text_encoder,
-                                 text_decoder,
+# image -> word
+decoder_trainer = DecoderTrainer(word_encoder,
+                                 word_decoder,
                                  decoder_gan,
                                  training_epochs=constants.TRAINING_EPOCHS,
                                  batch_size=constants.BATCH_SIZE)
@@ -193,26 +221,21 @@ for current_round in range(constants.TOTAL_TRAINING_ROUND):
 
     """ Train """
 
-    # text -> image
-    encoder_trainer.train(input_data=tokens_train,
+    # word -> image
+    encoder_trainer.train(input_data=word_embeddings_train,
                           exp_output_data=fashion_image_train_scaled)
 
-    # image -> text
-    decoder_trainer.train(input_data=tokens_train)
+    # image -> word
+    decoder_trainer.train(input_data=word_embeddings_train)
 
     """ Inference """
 
     # Select sample of sequences
     sample_indexes = np.random.randint(0,
-                                       tokens_test.shape[0],
+                                       word_embeddings_test.shape[0],
                                        constants.DISPLAY_ROW * constants.DISPLAY_COLUMN)
-    sample_tokens = tokens_test[sample_indexes]
-
-    sample_words = []
-    for token in sample_tokens:
-        word = index_word.get(token)
-        if word:
-            sample_words.append(word)
+    sample_words = words_test[sample_indexes]
+    sample_embeddings = word_embeddings_test[sample_indexes]
 
     # Select a random batch of images
     image_indexes = np.random.randint(0,
@@ -228,7 +251,7 @@ for current_round in range(constants.TOTAL_TRAINING_ROUND):
                                    should_save_to_file=should_save_to_file)
 
     # Encode images
-    encoded_sample_images_scaled = text_encoder.predict(sample_tokens)
+    encoded_sample_images_scaled = word_encoder.predict(sample_embeddings)
 
     # Evaluate
     loss_fake, acc_fake = encoder_discriminator.evaluate(encoded_sample_images_scaled,
@@ -241,7 +264,7 @@ for current_round in range(constants.TOTAL_TRAINING_ROUND):
     encoder_discriminator_loss.append(d_loss)
     encoder_discriminator_accuracy.append(d_acc)
 
-    g_loss, g_acc = encoder_gan.evaluate(sample_tokens, y_ones)
+    g_loss, g_acc = encoder_gan.evaluate(sample_embeddings, y_ones)
 
     encoder_generator_loss.append(g_loss)
     encoder_generator_accuracy.append(g_acc)
@@ -255,23 +278,35 @@ for current_round in range(constants.TOTAL_TRAINING_ROUND):
                                     should_display_directly=should_display_directly,
                                     should_save_to_file=should_save_to_file)
 
-    # Decode images to text
-    loss, accuracy = decoder_gan.evaluate(sample_tokens, sample_tokens)
+    # Decode images to word
+    loss, accuracy = decoder_gan.evaluate(sample_embeddings, sample_embeddings)
     decoder_loss.append(loss)
     decoder_accuracy.append(accuracy)
 
     print('Decode loss: {}, accuracy: {}'.format(loss, accuracy))
-    tokens_probs = decoder_gan.predict(sample_tokens)
-    decoded_words = []
-    for probs in tokens_probs:
-        token = np.argmax(probs)
-        word = index_word[token]
-        decoded_words.append(word)
+    decoded_embeddings = decoder_gan.predict(sample_embeddings)
 
+    decoded_words = []
+    for decoded_embedding in decoded_embeddings:
+        smallest_similarity = 1
+        nearest_word = None
+        for word in embeddings_index:
+            ref_embedding = embeddings_index.get(word)
+            if ref_embedding is not None:
+                # Calculate cosine similarity
+                similarity = cosine(decoded_embedding, ref_embedding)
+                if similarity < smallest_similarity:
+                    smallest_similarity = similarity
+                    nearest_word = word
+        if nearest_word is not None:
+            decoded_words.append(nearest_word)
+
+    print('Sample words:')
     print(sample_words)
+    print('Decoded words:')
     print(decoded_words)
 
-    # Display decoded text
+    # Display decoded word
     decoded_name = '{} - 3 - Decoded'.format(current_round + 1)
     text_displayer.display_samples(name=decoded_name,
                                    samples=decoded_words,
@@ -337,5 +372,7 @@ diagram_displayer.display_samples(name='Decoder Accuracy',
 
 print('Best accuracy: {}'.format(max(decoder_accuracy)))
 
-text_encoder.save('model/one_layer_text_fashion_encoder_generator.h5')
-text_decoder.save('model/one_layer_text_fashion_decoder_generator.h5')
+word_encoder.save(
+    'model/one_layer_word_fashion_embedding_encoder_generator.h5')
+word_decoder.save(
+    'model/one_layer_word_fashion_embedding_decoder_generator.h5')